refactor: extract more deserializers
