+++ src/deserializers/deserializeVersion.ts
@@ -1,0 +1,31 @@
+import { Version } from '../enums'
+
+export const deserializeVersion = (
+	line: string
+): {
+	version: Version
+	fullVersion: string
+} => {
+	let version = Version.Unsupported
+	const v = line.split('.')
+	const major = Number(v[0])
+	const minor = Number(v[1])
+
+	if (major <= 2) {
+		if (minor === 1) {
+			version = Version.v21x
+		} else if (minor === 2) {
+			version = Version.v22x
+		} else if (minor >= 3) {
+			// just parse anything newer as v2.3 as it's most likely closest
+			version = Version.v23x
+		}
+	} else {
+		version = Version.v23x
+	}
+
+	return {
+		version,
+		fullVersion: line,
+	}
+}
